Split out root.cc and role.cc, throw on bad timestamps
